+++ mayan/apps/documents/apps.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, unicode_literals
 
-from django.db.models.signals import post_delete
+from django.db.models.signals import post_delete, post_migrate
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.acls.classes import ModelPermission
@@ -43,8 +43,8 @@ from .events import (
     event_document_view
 )
 from .handlers import (
-    handler_create_default_document_type, handler_remove_empty_duplicates_lists,
-    handler_scan_duplicates_for,
+    handler_create_default_document_type, handler_create_document_cache,
+    handler_remove_empty_duplicates_lists, handler_scan_duplicates_for
 )
 from .links import (
     link_clear_image_cache, link_document_clear_transformations,
@@ -526,6 +526,10 @@ class DocumentsApp(MayanAppConfig):
         post_initial_setup.connect(
             dispatch_uid='handler_create_default_document_type',
             receiver=handler_create_default_document_type
+        )
+        post_migrate.connect(
+            dispatch_uid='documents_handler_create_document_cache',
+            receiver=handler_create_document_cache,
         )
         post_version_upload.connect(
             dispatch_uid='handler_scan_duplicates_for',